+++ event/preview.php
@@ -10,35 +10,24 @@
 
 namespace vse\TopicImagePreview\event;
 
-/**
- * @ignore
- */
-
 use phpbb\config\config;
 use phpbb\db\driver\driver_interface;
-use phpbb\language\language;
-use Symfony\Component\DomCrawler\Crawler;
+use phpbb\user;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
  * Topic Image Preview Event listener.
  */
-class listener implements EventSubscriberInterface
+class preview implements EventSubscriberInterface
 {
-	/**
-	 * @var \phpbb\config\config
-	 */
+	/** @var config */
 	protected $config;
 
-	/**
-	 * @var \phpbb\db\driver\driver_interface
-	 */
+	/** @var driver_interface */
 	protected $db;
 
-	/**
-	 * @var \phpbb\language\language
-	 */
-	protected $language;
+	/** @var user */
+	protected $user;
 
 	/**
 	 * {@inheritdoc}
@@ -46,8 +35,6 @@ class listener implements EventSubscriberInterface
 	public static function getSubscribedEvents()
 	{
 		return [
-			// ACP events
-			'core.acp_board_config_edit_add'		=> 'update_acp_data',
 			// Viewforum events
 			'core.viewforum_modify_topics_data'		=> 'update_row_data',
 			'core.viewforum_modify_topicrow'		=> 'update_tpl_data',
@@ -63,15 +50,15 @@ class listener implements EventSubscriberInterface
 	/**
 	 * Constructor
 	 *
-	 * @param \phpbb\config\config              $config
-	 * @param \phpbb\db\driver\driver_interface $db
-	 * @param \phpbb\language\language          $language
+	 * @param config           $config
+	 * @param driver_interface $db
+	 * @param user             $user
 	 */
-	public function __construct(config $config, driver_interface $db, language $language)
+	public function __construct(config $config, driver_interface $db, user $user)
 	{
 		$this->config = $config;
 		$this->db = $db;
-		$this->language = $language;
+		$this->user = $user;
 	}
 
 	/**
@@ -83,6 +70,11 @@ class listener implements EventSubscriberInterface
 	 */
 	public function update_row_data($event)
 	{
+		if (empty($this->user->data['user_vse_tip']))
+		{
+			return;
+		}
+
 		// Use topic_list from event, otherwise create one based on the rowset
 		$topic_list = $event->offsetExists('topic_list') ? $event['topic_list'] : array_keys($event['rowset']);
 
@@ -90,6 +82,26 @@ class listener implements EventSubscriberInterface
 		{
 			$event['rowset'] = $this->query_images($topic_list, $event['rowset']);
 		}
+	}
+
+	/**
+	 * Add image previews to the template data.
+	 *
+	 * @param \phpbb\event\data $event The event object
+	 *
+	 * @return void
+	 */
+	public function update_tpl_data($event)
+	{
+		// Check if we have any post text or images
+		if (empty($this->user->data['user_vse_tip']) || empty($event['row']['post_text']) || !preg_match('/^<[r][ >]/', $event['row']['post_text']) || strpos($event['row']['post_text'], '<IMG ') === false)
+		{
+			return;
+		}
+
+		// Send the image string to the template
+		$block = $event->offsetExists('topic_row') ? 'topic_row' : 'tpl_ary';
+		$event[$block] = array_merge($event[$block], ['TOPIC_IMAGES' => $this->extract_images($event['row']['post_text'])]);
 	}
 
 	/**
@@ -136,73 +148,27 @@ class listener implements EventSubscriberInterface
 	}
 
 	/**
-	 * Add image previews to the template data.
+	 * Extract images from a post and return them as HTML image tags.
 	 *
-	 * @param \phpbb\event\data $event The event object
+	 * @param string $post Post text from the database.
 	 *
-	 * @return void
+	 * @return string An string of HTML IMG tags.
 	 */
-	public function update_tpl_data($event)
+	protected function extract_images($post)
 	{
-		// Check if we have any post text or images
-		$row = $event['row'];
-		if (empty($row['post_text']) || !preg_match('/^<[r][ >]/', $row['post_text']) || strpos($row['post_text'], '<IMG ') === false)
-		{
-			return;
-		}
-
 		// Extract the images
-		$crawler = new Crawler($row['post_text']);
-		$images = $crawler->filterXpath('//img[not(ancestor::img)]')->extract(['src']);
+		$images = [];
+		$dom = new \DOMDocument;
+		$dom->loadXML($post);
+		$xpath = new \DOMXPath($dom);
+		foreach ($xpath->query('//IMG[not(ancestor::IMG)]/@src') as $image)
+		{
+			$images[] = $image->textContent;
+		}
 
 		// Create a string of images
-		$img_string = implode(' ', array_map(function ($image) {
+		return implode(' ', array_map(function ($image) {
 			return "<img src='{$image}' alt='' style='max-width:{$this->config['vse_tip_dim']}px; max-height:{$this->config['vse_tip_dim']}px;' />";
 		}, array_slice($images, 0, (int) $this->config['vse_tip_num'], true)));
-
-		// Send the image string to the template
-		$block = $event->offsetExists('topic_row') ? 'topic_row' : 'tpl_ary';
-		$event[$block] = array_merge($event[$block], ['TOPIC_IMAGES' => $img_string]);
-	}
-
-	/**
-	 * Add ACP config options to Post settings.
-	 *
-	 * @param \phpbb\event\data $event The event object
-	 *
-	 * @return void
-	 */
-	public function update_acp_data($event)
-	{
-		$display_vars = $event['display_vars'];
-		if ($event['mode'] === 'post' && array_key_exists('legend3', $display_vars['vars']))
-		{
-			$this->language->add_lang('tip_acp', 'vse/TopicImagePreview');
-
-			$my_config_vars = [
-				'legend_vse_tip'	=> 'ACP_TIP_TITLE',
-				'vse_tip_new'		=> ['lang' => 'ACP_TIP_DISPLAY_AGE', 'validate' => 'bool', 'type' => 'custom', 'function' => [$this, 'select_vse_tip_new'], 'explain' => true],
-				'vse_tip_num'		=> ['lang' => 'ACP_TIP_DISPLAY_NUM', 'validate' => 'int:0:99', 'type' => 'number:0:99', 'explain' => true],
-				'vse_tip_dim'		=> ['lang' => 'ACP_TIP_DISPLAY_DIM', 'validate' => 'int:0:999', 'type' => 'number:0:999', 'explain' => true, 'append' => ' ' . $this->language->lang('PIXEL')],
-			];
-
-			$display_vars['vars'] = phpbb_insert_config_array($display_vars['vars'], $my_config_vars, ['before' => 'legend3']);
-			$event['display_vars'] = $display_vars;
-		}
-	}
-
-	/**
-	 * Create custom radio buttons.
-	 *
-	 * @param mixed  $value
-	 * @param string $key
-	 *
-	 * @return string
-	 */
-	public function select_vse_tip_new($value, $key = '')
-	{
-		$radio_ary = [1 => 'ACP_TIP_NEWEST_POST', 0 => 'ACP_TIP_OLDEST_POST'];
-
-		return h_radio('config[vse_tip_new]', $radio_ary, $value, $key);
 	}
 }