+++ migrations/ucp_option.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ *
+ * Topic Image Preview. An extension for the phpBB Forum Software package.
+ *
+ * @copyright (c) 2019, Matt Friedman
+ * @license GNU General Public License, version 2 (GPL-2.0)
+ *
+ */
+
+namespace vse\TopicImagePreview\migrations;
+
+class ucp_option extends \phpbb\db\migration\migration
+{
+	public function effectively_installed()
+	{
+		return $this->db_tools->sql_column_exists($this->table_prefix . 'users', 'user_vse_tip');
+	}
+
+	public static function depends_on()
+	{
+		return ['\vse\TopicImagePreview\migrations\install_tip'];
+	}
+
+	public function update_schema()
+	{
+		return [
+			'add_columns'	=> [
+				$this->table_prefix . 'users'	=> [
+					'user_vse_tip'	=> ['BOOL', 1],
+				],
+			],
+		];
+	}
+
+	public function revert_schema()
+	{
+		return [
+			'drop_columns'	=> [
+				$this->table_prefix . 'users'	=> [
+					'user_vse_tip',
+				],
+			],
+		];
+	}
+}