+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0-b5 - 2019-06-19
+
+- Added a UCP display preference so users can disable image previews.
+- Reverted back from Symfony's crawler to PHP's DOMDocument crawler.
+
 ## 1.0.0-b4 - 2019-05-22
 
 - Switch to Symfony's crawler to parse images out of posts.